Create a seperate component for the navbar.
